services/api/requirements.txt
# Web framework
fastapi==0.115.4
uvicorn[standard]==0.32.1
python-multipart==0.0.12

# Database
sqlalchemy==2.0.36
asyncpg==0.30.0
alembic==1.14.0
pgvector==0.3.5

# Auth (Supabase)
supabase==2.9.1
python-jose[cryptography]==3.3.0

# AI / NLP
sentence-transformers==3.3.1
spacy==3.8.3
transformers==4.46.3
torch==2.5.1
numpy==1.26.4
scikit-learn==1.5.2

# PDF processing
pymupdf==1.24.13
pdfplumber==0.11.4

# Task queue
celery==5.4.0
redis==5.2.0
flower==2.0.1

# Validation / serialization
pydantic==2.10.1
pydantic-settings==2.6.1
email-validator==2.2.0

# HTTP client
httpx==0.28.0

# Utilities
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
tenacity==9.0.0
loguru==0.7.2

# Dev / test
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.28.0
factory-boy==3.3.1
